Changed: use fmt-config instead of finding fmt. Fixed: added content mismatch with wchar_t.

+++ include/common613/file_utils.h
@@ -61,7 +61,7 @@ struct WcstombsWrapper{
 
 template <>
 struct WcstombsWrapper<wchar_t>{
-  inline static size_t call(char* buffer, const wchar_t* path, int bufferSize) {
+  inline static size_t call(char* buffer, const wchar_t* path, size_t bufferSize) {
     return std::wcstombs(buffer, path, bufferSize);
   }
 };
